Issue 3: For create_tables_from_dict, use default of not clobbering dataset.
This also revealed an inconsistency in the function tables(), where bq returned only the table names, but mock_bq returned the full table paths. I made mock_bq return the table names and made adjustments as needed to fit that.

Also change the behavior of create_tables_from_dict to raise exceptions
if it finds tables it's supposed to be creating.

Change-Id: I49145f093682df8c8b347c316b743603e949fbc6