Change behavior of mock_bq create_dataset_by_name
The existing behavior caused mock_bq to delete a dataset (though not its
constituent tables) when creating a dataset that already existed.  This
leaves the data in an inconsistent state (the dataset is gone but its
constituent tables are not, so e.g. a follow-on create_tables_from_dict
will not behave correctly with replace_existing_tables=True and delete
new tables since it doesn't know they exist).

More crucially, this behavior is not the same as real bq, which just
issues a warning rather than deleting the dataset and starting fresh.

This change makes mock bq match real bq in functionality.

Change-Id: Iab37e650b9c67d89ba0253a8fa0a21a8bc4dd629